move parameter ui utils to separate file
