+++ frontend/src/metabase/parameters/components/ParametersList.jsx
@@ -9,7 +9,7 @@ import {
   SortableElement,
   SortableHandle,
 } from "metabase/components/sortable";
-import { getVisibleParameters } from "metabase/meta/Parameter";
+import { getVisibleParameters } from "metabase/parameters/utils/ui";
 import { getValuePopulatedParameters } from "metabase/parameters/utils/parameter-values";
 
 import type {
