+++ frontend/test/metabase/meta/Parameter.unit.spec.js
@@ -1,9 +1,5 @@
 import MetabaseSettings from "metabase/lib/settings";
-import {
-  getTemplateTagParameters,
-  buildHiddenParametersSlugSet,
-  getVisibleParameters,
-} from "metabase/meta/Parameter";
+import { getTemplateTagParameters } from "metabase/meta/Parameter";
 
 MetabaseSettings.get = jest.fn();
 
@@ -172,55 +168,6 @@ describe("metabase/meta/Parameter", () => {
         ];
         expect(getTemplateTagParameters(tags)).toEqual(parameters);
       });
-    });
-  });
-
-  describe("buildHiddenParametersSlugSet", () => {
-    it("should turn the given string of slugs separated by commas into a set of slug strings", () => {
-      expect(buildHiddenParametersSlugSet("a,b,c")).toEqual(
-        new Set(["a", "b", "c"]),
-      );
-    });
-
-    it("should return an empty set for any input that is not a string", () => {
-      expect(buildHiddenParametersSlugSet(undefined)).toEqual(new Set());
-      expect(buildHiddenParametersSlugSet(111111)).toEqual(new Set());
-    });
-  });
-
-  describe("getVisibleParameters", () => {
-    const parameters = [
-      {
-        id: 1,
-        slug: "foo",
-      },
-      {
-        id: 2,
-        slug: "bar",
-      },
-      {
-        id: 3,
-        slug: "baz",
-      },
-      {
-        id: 4,
-        slug: "qux",
-      },
-    ];
-
-    const hiddenParameterSlugs = "bar,baz";
-
-    it("should return the parameters that are not hidden", () => {
-      expect(getVisibleParameters(parameters, hiddenParameterSlugs)).toEqual([
-        {
-          id: 1,
-          slug: "foo",
-        },
-        {
-          id: 4,
-          slug: "qux",
-        },
-      ]);
     });
   });
 });
